BoardService refactor v1
