--- service/src/main/java/hu/unideb/inf/rft/neuban/service/handler/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author Erdei Krisztián
- *
- */
-package hu.unideb.inf.rft.neuban.service.handler;
