+++ service/src/main/java/hu/unideb/inf/rft/neuban/service/interfaces/BoardService.java
@@ -1,16 +1,28 @@
 package hu.unideb.inf.rft.neuban.service.interfaces;
 
-import hu.unideb.inf.rft.neuban.service.domain.BoardDto;
-import hu.unideb.inf.rft.neuban.service.exceptions.BoardNotFoundException;
-
 import java.util.List;
 import java.util.Optional;
 
+import hu.unideb.inf.rft.neuban.service.domain.BoardDto;
+import hu.unideb.inf.rft.neuban.service.exceptions.BoardNotFoundException;
+import hu.unideb.inf.rft.neuban.service.exceptions.NonExistentBoardIdException;
+import hu.unideb.inf.rft.neuban.service.exceptions.NonExistentUserIdException;
+import hu.unideb.inf.rft.neuban.service.exceptions.RelationNotFoundException;
+
 public interface BoardService {
 
-    Optional<BoardDto> get(Long boardId);
+	Optional<BoardDto> get(Long boardId);
 
-    List<BoardDto> getAllByUserId(Long userId);
+	List<BoardDto> getAllByUserId(Long userId);
 
-    Long update(BoardDto boardDto) throws BoardNotFoundException;
+	Long update(BoardDto boardDto) throws BoardNotFoundException;
+
+	void removeUserFromBoardByUserIdAndByBoardId(Long userId, Long boardId)
+			throws NonExistentBoardIdException, RelationNotFoundException, NonExistentUserIdException;
+
+	void createBoardByDefaultUserIdAndByTitle(Long userId, String title) throws NonExistentUserIdException;
+
+	void addUserToBoardByUserIdAndByBoardId(Long userId, Long boardId)
+			throws NonExistentBoardIdException, NonExistentUserIdException;
+
 }
